refactor: 함수분리
